rename package on io.singularitynet.sdk.tutorial
