+++ example/android-tutorial/img-segm-demo-app/app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 29
 
     defaultConfig {
-        applicationId "io.singularitynet.sdk.tutorial"
+        applicationId "io.singularitynet.sdk.tutorial.semanticsegmentation"
         minSdkVersion 24
         targetSdkVersion 29
         versionCode 1
