+++ example/android-tutorial/img-segm-demo-app/semantic-segmentation-api/src/androidTest/java/io/singularitynet/sdk/tutorial/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.semantic_segmentation_api;
+package io.singularitynet.sdk.tutorial;
 
 import android.content.Context;
