+++ example/android-tutorial/img-segm-demo-app/semantic-segmentation-api/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.semantic_segmentation_api">
+    package="io.singularitynet.sdk.tutorial">
 
     /
 </manifest>
