+++ example/android-tutorial/img-segm-demo-app/semantic-segmentation-api/src/test/java/io/singularitynet/sdk/tutorial/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.semantic_segmentation_api;
+package io.singularitynet.sdk.tutorial;
 
 import org.junit.Test;
